requirements.txt
# Core dependencies
flask
python-dotenv

# Hyperliquid dependencies
hyperliquid-python-sdk
eth-account

# Database dependencies
pymongo

# Other dependencies
requests
typing-extensions